+++ packages/contracts/src/plugins/governance/majority-voting/token/TokenVoting.sol
@@ -2,7 +2,8 @@
 
 pragma solidity 0.8.17;
 
-import {IVotesUpgradeable} from "@openzeppelin/contracts-upgradeable/governance/utils/IVotesUpgradeable.sol";
+import {IERC5805Upgradeable, IVotesUpgradeable} from "@openzeppelin/contracts-upgradeable/interfaces/IERC5805Upgradeable.sol";
+import {IERC6372Upgradeable} from "@openzeppelin/contracts-upgradeable/interfaces/IERC6372Upgradeable.sol";
 import {SafeCastUpgradeable} from "@openzeppelin/contracts-upgradeable/utils/math/SafeCastUpgradeable.sol";
 import {IERC20Upgradeable} from "@openzeppelin/contracts-upgradeable/token/ERC20/IERC20Upgradeable.sol";
 
@@ -16,7 +17,7 @@ import {IMajorityVoting} from "../IMajorityVoting.sol";
 /// @author Aragon Association - 2021-2023
 /// @notice The majority voting implementation using an [OpenZeppelin `Votes`](https://docs.openzeppelin.com/contracts/4.x/api/governance#Votes) compatible governance token.
 /// @dev This contract inherits from `MajorityVotingBase` and implements the `IMajorityVoting` interface.
-contract TokenVoting is IMembership, MajorityVotingBase {
+contract TokenVoting is IMembership, IERC6372Upgradeable, MajorityVotingBase {
     using SafeCastUpgradeable for uint256;
 
     /// @notice The [ERC-165](https://eips.ethereum.org/EIPS/eip-165) interface ID of the contract.
@@ -24,7 +25,7 @@ contract TokenVoting is IMembership, MajorityVotingBase {
         this.initialize.selector ^ this.getVotingToken.selector;
 
     /// @notice An [OpenZeppelin `Votes`](https://docs.openzeppelin.com/contracts/4.x/api/governance#Votes) compatible contract referencing the token being used for voting.
-    IVotesUpgradeable private votingToken;
+    IERC5805Upgradeable private votingToken;
 
     /// @notice Thrown if the voting power is zero
     error NoVotingPower();
@@ -41,7 +42,7 @@ contract TokenVoting is IMembership, MajorityVotingBase {
     ) external initializer {
         __MajorityVotingBase_init(_dao, _votingSettings);
 
-        votingToken = _token;
+        votingToken = IERC5805Upgradeable(address(_token));
 
         emit MembershipContractAnnounced({definingContract: address(_token)});
     }
@@ -64,8 +65,28 @@ contract TokenVoting is IMembership, MajorityVotingBase {
     }
 
     /// @inheritdoc MajorityVotingBase
-    function totalVotingPower(uint256 _blockNumber) public view override returns (uint256) {
-        return votingToken.getPastTotalSupply(_blockNumber);
+    function totalVotingPower(uint256 _timepoint) public view override returns (uint256) {
+        return votingToken.getPastTotalSupply(_timepoint);
+    }
+
+    /// @dev Clock (as specified in EIP-6372) is set to match the token's clock. Fallback to block numbers if the token
+    /// does not implement EIP-6372.
+    function clock() public view virtual override returns (uint48) {
+        try votingToken.clock() returns (uint48 timepoint) {
+            return timepoint;
+        } catch {
+            return block.number.toUint48();
+        }
+    }
+
+    /// @dev Machine-readable description of the clock as specified in EIP-6372.
+    /// solhint-disable-next-line func-name-mixedcase
+    function CLOCK_MODE() public view virtual override returns (string memory) {
+        try votingToken.CLOCK_MODE() returns (string memory clockmode) {
+            return clockmode;
+        } catch {
+            return "mode=blocknumber&from=default";
+        }
     }
 
     /// @inheritdoc MajorityVotingBase
@@ -94,12 +115,9 @@ contract TokenVoting is IMembership, MajorityVotingBase {
             }
         }
 
-        uint256 snapshotBlock;
-        unchecked {
-            snapshotBlock = block.number - 1; // The snapshot block must be mined already to protect the transaction against backrunning transactions causing census changes.
-        }
+        uint48 snapshotTimepoint = clock() - 1; // The snapshot timepoint must be mined already to protect the transaction against backrunning transactions causing census changes.
 
-        uint256 totalVotingPower_ = totalVotingPower(snapshotBlock);
+        uint256 totalVotingPower_ = totalVotingPower(snapshotTimepoint);
 
         if (totalVotingPower_ == 0) {
             revert NoVotingPower();
@@ -121,7 +139,8 @@ contract TokenVoting is IMembership, MajorityVotingBase {
 
         proposal_.parameters.startDate = _startDate;
         proposal_.parameters.endDate = _endDate;
-        proposal_.parameters.snapshotBlock = snapshotBlock.toUint64();
+        proposal_.parameters.snapshotBlock = (block.number - 1).toUint64();
+        proposal_.parameters.snapshotTimepoint = snapshotTimepoint;
         proposal_.parameters.votingMode = votingMode();
         proposal_.parameters.supportThreshold = supportThreshold();
         proposal_.parameters.minVotingPower = _applyRatioCeiled(
@@ -164,7 +183,7 @@ contract TokenVoting is IMembership, MajorityVotingBase {
         Proposal storage proposal_ = proposals[_proposalId];
 
         // This could re-enter, though we can assume the governance token is not malicious
-        uint256 votingPower = votingToken.getPastVotes(_voter, proposal_.parameters.snapshotBlock);
+        uint256 votingPower = votingToken.getPastVotes(_voter, proposal_.parameters.snapshotTimepoint);
         VoteOption state = proposal_.voters[_voter];
 
         // If voter had previously voted, decrease count
@@ -218,7 +237,7 @@ contract TokenVoting is IMembership, MajorityVotingBase {
         }
 
         // The voter has no voting power.
-        if (votingToken.getPastVotes(_account, proposal_.parameters.snapshotBlock) == 0) {
+        if (votingToken.getPastVotes(_account, proposal_.parameters.snapshotTimepoint) == 0) {
             return false;
         }
 
@@ -231,6 +250,23 @@ contract TokenVoting is IMembership, MajorityVotingBase {
         }
 
         return true;
+    }
+
+    /// @inheritdoc MajorityVotingBase
+    function isSupportThresholdReachedEarly(
+        uint256 _proposalId
+    ) public view virtual override returns (bool) {
+        Proposal storage proposal_ = proposals[_proposalId];
+
+        uint256 noVotesWorstCase = totalVotingPower(proposal_.parameters.snapshotTimepoint) -
+            proposal_.tally.yes -
+            proposal_.tally.abstain;
+
+        // The code below implements the formula of the early execution support criterion explained in the top of this file.
+        // `(1 - supportThreshold) * N_yes > supportThreshold *  N_no,worst-case`
+        return
+            (RATIO_BASE - proposal_.parameters.supportThreshold) * proposal_.tally.yes >
+            proposal_.parameters.supportThreshold * noVotesWorstCase;
     }
 
     /// @dev This empty reserved space is put in place to allow future versions to add new