+++ packages/contracts/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a `nonReentrant` modifier to the `execute` function in the `DAO` contract.
 - Added `allowFailureMap` to `IDAO.Executed` event.
 - Added `initializeFrom()` reinitializer to `DAO` contract.
+- Support ERC-6372 in TokenVoting plugin.
 
 ### Changed